requirements.txt
python-dotenv
langchain
langchain_openai
langchain_community
langchain_chroma
pypdf
gradio